requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
Django==4.1.6
django-environ==0.9.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==1.7.1
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
